pyproject.toml
[project]
name = "toss"
version = "0.2.0"
description = "Toss: Agent-to-Agent document sharing tool"
readme = "README.md"
authors = [
    { name = "韩昊珂(Han Haoke)", email = "hanhaoke@qdu.edu.cn" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
    # T2: libsodium bindings (X25519 + Ed25519 + SecretBox) for client-side
    # end-to-end encryption. Ships as a wheel on all supported platforms.
    "pynacl>=1.5",
]

[project.scripts]
toss = "toss.cli.main:main"
toss-mcp = "toss.mcp.server:main"

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
    "mypy>=1.10",
    "respx>=0.22",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.13"
strict = true
